Move login tests to separate file
